+++ tests/extendrtests/R/extendr-wrappers.R
@@ -14,6 +14,12 @@ hello_world <- function() .Call(wrap__hello_world)
 
 do_nothing <- function() invisible(.Call(wrap__do_nothing))
 
+#' This is invisible by default
+result_unit <- function() invisible(.Call(wrap__result_unit))
+
+#' Return a string but invisibly
+invisible_string <- function() invisible(.Call(wrap__invisible_string))
+
 double_scalar <- function(x) .Call(wrap__double_scalar, x)
 
 int_scalar <- function(x) .Call(wrap__int_scalar, x)