+++ tests/extendrtests/src/rust/src/lib.rs
@@ -27,6 +27,18 @@ fn hello_world() -> &'static str {
 #[extendr]
 fn do_nothing() {}
 
+/// This is invisible by default
+#[extendr]
+fn result_unit() -> Result<()> {
+    Ok(())
+}
+
+/// Return a string but invisibly
+#[extendr(invisible)]
+fn invisible_string() -> &'static str {
+    "This should be invisible"
+}
+
 // TryFrom: conversions
 
 // Atomic types
@@ -308,6 +320,8 @@ extendr_module! {
     mod extendrtests;
     fn hello_world;
     fn do_nothing;
+    fn result_unit;
+    fn invisible_string;
 
     fn double_scalar;
     fn int_scalar;